argus_track/requirements.txt
# argus_track/requirements.txt (UPDATED WITH GPS EXTRACTION)

# Core dependencies
numpy>=1.19.0
scipy>=1.5.0
opencv-python>=4.5.0
filterpy>=1.4.5
numba>=0.53.0  # For JIT compilation

# PyTorch for YOLOv11 support
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# YOLOv11 specific
ultralytics>=8.0.0  # For YOLOv11 support

# Optional optimizations
lap>=0.4.0  # Faster Hungarian algorithm

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Development dependencies
pytest>=6.0.0
pytest-benchmark>=3.4.0
black>=21.0
flake8>=3.9.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# GPS support
pyproj>=3.0.0  # For GPS coordinate transformations
scikit-learn>=0.24.0  # For clustering in static analysis
pynvml>=11.0.0  # Optional: For GPU monitoring

# GPS visualization
folium>=0.12.0  # For interactive maps
geojson>=2.5.0  # For GeoJSON export

# Stereo vision and 3D processing
transforms3d>=0.3.1  # For 3D transformations
open3d>=0.13.0  # Optional: For 3D visualization

# Configuration and data handling
pyyaml>=5.4.0  # For YAML configuration files
pandas>=1.3.0  # For data analysis and CSV handling

# Image processing enhancements
Pillow>=8.0.0  # Enhanced image processing
scikit-image>=0.18.0  # Additional image processing algorithms

# GPS EXTRACTION DEPENDENCIES
# ============================

# HTML/XML parsing for GPS metadata
beautifulsoup4>=4.9.0  # For parsing GPS metadata from ExifTool output
lxml>=4.6.0  # XML parser backend for BeautifulSoup

# GPS metadata handling
GPSPhoto>=2.2.0  # For reading/writing GPS metadata in images (optional)

# GoPro telemetry extraction (optional but recommended)
gopro-overlay>=0.10.0  # For GoPro telemetry data extraction

# ExifTool integration (ExifTool must be installed separately)
# Note: ExifTool (https://exiftool.org/) must be installed on the system
# Windows: Download from https://exiftool.org/
# macOS: brew install exiftool
# Linux: sudo apt-get install libimage-exiftool-perl

# Video processing enhancements
ffmpeg-python>=0.2.0  # For video metadata extraction (alternative method)

# Time and date handling
python-dateutil>=2.8.0  # Enhanced date parsing

# Process management
psutil>=5.8.0  # For system monitoring during processing